test(control-ui): extract the shared render setup out of the duplicated specs
The ControlUI specs each carried their own copy of the mount/unmount
container lifecycle, the icon module stubs and the fake connection, view
and stream builders. Move that setup into src/testHelpers.tsx (plus a
dependency-free src/testIconStubs.tsx that the vi.mock factories can pull
in without dragging the component tree into the factory) and rewrite the
specs on top of it. The assertions are unchanged.
