+++ packages/streamwall-control-ui/src/hotkeysOptions.test.tsx
@@ -1,88 +1,31 @@
-import { render } from 'preact'
-import { act } from 'preact/test-utils'
-import type { StreamDelayStatus } from 'streamwall-shared'
 import { afterEach, describe, expect, test, vi } from 'vitest'
-import * as Y from 'yjs'
-import type { StreamwallConnection } from './index.tsx'
-import { ControlUI } from './index.tsx'
+import { makeConnection, renderControlUI } from './testHelpers.tsx'
 
-// react-icons renders through preact/compat's Context.Consumer, which
-// currently crashes under this package's happy-dom test environment
-// (unrelated to the hotkey wiring under test here) - stub the icons out so
-// the component can render far enough to register its hotkeys.
-vi.mock('react-icons/fa', () => ({
-  FaExchangeAlt: () => null,
-  FaExclamationTriangle: () => null,
-  FaRedoAlt: () => null,
-  FaRegLifeRing: () => null,
-  FaRegWindowMaximize: () => null,
-  FaSyncAlt: () => null,
-  FaVideoSlash: () => null,
-  FaVolumeUp: () => null,
-}))
-vi.mock('react-icons/md', () => ({
-  MdOutlineStayCurrentLandscape: () => null,
-  MdOutlineStayCurrentPortrait: () => null,
-}))
+vi.mock(
+  'react-icons/fa',
+  async () => (await import('./testIconStubs.tsx')).faIconStubs,
+)
+vi.mock(
+  'react-icons/md',
+  async () => (await import('./testIconStubs.tsx')).mdIconStubs,
+)
 
 const useHotkeysMock = vi.hoisted(() => vi.fn())
 vi.mock('react-hotkeys-hook', () => ({
   useHotkeys: useHotkeysMock,
 }))
 
-let container: HTMLDivElement | undefined
-
 afterEach(() => {
-  if (container) {
-    act(() => render(null, container!))
-    container.remove()
-    container = undefined
-  }
   useHotkeysMock.mockClear()
 })
 
-function renderControlUI(): HTMLDivElement {
-  container = document.createElement('div')
-  document.body.appendChild(container)
-
-  const delayState: StreamDelayStatus = {
-    isConnected: true,
-    delaySeconds: 0,
-    restartSeconds: 0,
-    isCensored: false,
-    isStreamRunning: true,
-    startTime: 0,
-    state: 'idle',
-  }
-
-  const connection: StreamwallConnection = {
-    isConnected: true,
-    role: 'operator',
-    send: () => {},
-    sharedState: { views: {} },
-    stateDoc: new Y.Doc(),
-    config: undefined,
-    streams: [],
-    customStreams: [],
-    views: [],
-    fullscreenViewIdx: null,
-    stateIdxMap: new Map(),
-    delayState,
-    authState: undefined,
-    layoutPresets: [],
-    favorites: [],
-    dataSourceHealth: [],
-  }
-
-  act(() => {
-    render(<ControlUI connection={connection} />, container!)
-  })
-  return container
+function renderWithHotkeys(): HTMLDivElement {
+  return renderControlUI(makeConnection())
 }
 
 describe('alt+<n> listen-toggle hotkey', () => {
   test('enables the hotkey while a grid input is focused via the v5 enableOnFormTags option', () => {
-    renderControlUI()
+    renderWithHotkeys()
 
     const listenToggleCall = useHotkeysMock.mock.calls.find(
       ([keys]) =>
@@ -105,7 +48,7 @@ describe('alt+<n> listen-toggle hotkey', () => {
 
 describe('alt+ctrl+<n> second-layer listen-toggle hotkey (#240)', () => {
   test('registers an alt+ctrl chord layer covering the same 20 trigger keys', () => {
-    renderControlUI()
+    renderWithHotkeys()
 
     const secondLayerCall = useHotkeysMock.mock.calls.find(
       ([keys]) =>
@@ -124,7 +67,7 @@ describe('alt+ctrl+<n> second-layer listen-toggle hotkey (#240)', () => {
 
 describe('alt+shift+<n> blur-toggle hotkey', () => {
   test('registers a base alt+shift chord layer covering 20 trigger keys', () => {
-    renderControlUI()
+    renderWithHotkeys()
 
     const baseLayerCall = useHotkeysMock.mock.calls.find(
       ([keys]) =>
@@ -142,7 +85,7 @@ describe('alt+shift+<n> blur-toggle hotkey', () => {
 
 describe('alt+ctrl+shift+<n> second-layer blur-toggle hotkey (#294)', () => {
   test('registers an alt+ctrl+shift chord layer covering the same 20 trigger keys', () => {
-    renderControlUI()
+    renderWithHotkeys()
 
     const secondLayerCall = useHotkeysMock.mock.calls.find(
       ([keys]) =>
